+++ libIRDB/include/core/basetypes.hpp
@@ -18,8 +18,8 @@
  *
  */
 
+#include <stdint.h>
 
-
-typedef int virtual_offset_t;
+typedef uintptr_t virtual_offset_t;
 typedef int db_id_t;
 typedef int schema_version_t;
